Rename amb to cmp
Traditional amb effect has a different semantics and we cannot achieve it using current Ruby means. It's possible with `callcc` but it was deprecated a long time ago. This commit renames current effect to `cmp` which is also 3-letters long but  acts a bit differently. It's still very useful for demonstration purposes